py/libs/agentkit/pyproject.toml
[project]
name = "agentkit"
version = "0.1.0"
description = "YAML-driven agent configuration helpers"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.0",
  "jinja2>=3.1.4",
  "pydantic>=2.10.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
